Fix problems with duplicates and empties
Boris Novikov pointed out that my semijoin SQL only works when you're
doing an equijoin between an FK on the left and a PK on the right. But
otherwise you need to avoid overlapping duplicates for each left-hand
record.

I also found some problems with empties (a perennial issue it seems!).
For antijoins, we need to handle those outside the outer join (and
before unnesting), so we can distinguish between "didn't find a match"
and "found a match that erased the entire `a.valid_at`".